refactor some logic to engine
